feat: notes adapter example in wirespec

+++ domain/src/main/kotlin/community/flock/pragmatic/domain/pet/PetAdapter.kt
@@ -3,7 +3,7 @@ package community.flock.pragmatic.domain.pet
 import community.flock.pragmatic.domain.pet.model.Pet
 
 interface PetAdapter {
-    suspend fun getPetById(id: Int): Pet
+    suspend fun getPetById(id: Long): Pet
 
     suspend fun savePet(pet: Pet): Pet
 }
